elang/compiler/semantics: Make |sm::ArrayType::rank()| to return |size_t| for avoiding |static_cast<int>()|.

+++ elang/compiler/semantics/nodes.h
@@ -147,7 +147,7 @@ class ArrayType final : public Type {
   // Dimension of each rank. dimensions.front() == -1 means unbound array.
   const ZoneVector<int>& dimensions() const { return dimensions_; }
   Type* element_type() const { return element_type_; }
-  int rank() const { return static_cast<int>(dimensions_.size()); }
+  size_t rank() const { return dimensions_.size(); }
 
  private:
   ArrayType(Zone* zone, Type* element_type, const std::vector<int>& dimensions);
@@ -280,7 +280,7 @@ class EnumMember final : public NamedMember<Semantic> {
 //
 // Field
 //
-class Field final : public NamedMember<Semantic>,  public WithModifiers {
+class Field final : public NamedMember<Semantic>, public WithModifiers {
   DECLARE_CONCRETE_SEMANTIC_CLASS(Field, Semantic);
 
  public:
@@ -334,7 +334,7 @@ class Literal final : public Value {
 //
 // Method
 //
-class Method final : public NamedMember<Semantic>,  public WithModifiers {
+class Method final : public NamedMember<Semantic>, public WithModifiers {
   DECLARE_CONCRETE_SEMANTIC_CLASS(Method, Semantic);
 
  public:
